requirements.txt
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.0
pyyaml==6.0.1 